refactor the existing mdlw
