+++ src/auth/mod.rs
@@ -1,3 +1,2 @@
 pub mod authentication;
 pub mod errors;
-pub mod middleware;
